requirements.txt
Flask_WTF==0.14.3
pandas==1.1.5
six==1.12.0
tensorflow==2.2.0
WTForms_Components==0.10.5
stripe==2.27.0
deepface==0.0.49
Werkzeug==1.0.1
opencv_contrib_python==4.5.2.52
WTForms==2.3.3
Flask_MySQLdb==0.2.0
Flask==1.1.2
Flask_Mail==0.9.1
matplotlib==3.3.3
Flask_Cors==3.0.10
numpy==1.17.3
wget==3.2
python_docx==0.8.10
nltk==3.6.2
coolname==1.1.0
object_detection==0.0.3
Pillow==8.3.2
Jinja2==3.0.3
itsdangerous==2.0.1
protobuf==3.20.3
Flask-Session==0.3.2
dlib==19.24.4
bcrypt==3.2.0
requests==2.31
python-dotenv ==1.0.1
Flask-S3==0.3.3
boto3==1.18.29